Added AsyncSignal
This class operates very similarly to the scheduler, but is much more lightweight.
Unlike the scheduler, this can be used in a similar manner to a condition variable
where you can observe some changes, and then wait for further changes to the observable,
but never having the race condition that is present if the IScheduler is used this way.

For example, the following code has a race condition bug where the read
and the block are not atomic, and so if the state changes between the state read
and the yield, the state update will be missed. In this case, the race condition
window has been artificially increased to ~500ms:

```cs
while (true)
{
    StateCopy = Something.State;
    await Task.Delay(500);
    await Something.Scheduler.Yield();
}
```

But with the AsyncSignal, we can correctly handle this behavior and receive the
new state update too by calling the Wait() before we read the state, and then
later on. If the signal has been fired, then the subsequent `await waitTask`
will be a no op:

```cs
while (true)
{
    var waitTask = Something.Signal.Wait();
    StateCopy = Something.State;
    await Task.Delay(500);
    await waitTask;
}
```